Add afex_aov example dependency
The example in the tidy_stats.afex_aov documentation was not
running. It depends on the afex package. This commit makes the
example throw a clear error that the required afex package is
missing when the example is run. It also adds the afex package to
the suggested packages for developer convienience.