Add emmGrid example dependency
The example in the tidy_stats.emmGrid documentation was not
running. It depends on the emmeans package. This commit makes the
example throw a clear error that the required emmeans package is
missing when the example is run. It also adds the emmeans package to
the suggested packages for developer convienience.