Add tidy_stats.mixed example dependency
The example in the tidy_stats.mixed documentation was not running. It
depends on the afex package. This commit makes the example throw a
clear error that the required afex package is missing when the example
is run.

NOTE: The example takes very long to run. Something may need to be
done about this.