Make examples friendlier to testing
This commit makes it possible to test this package when not all
suggested packages are installed. Suggested packages in DESCRIPTION
are not automatically installed. These packages are often required for
examples. By encapsulating the examples in requireNamespace("pkg",
quietly = TRUE) conditionals, we can make sure the examples only run
when the required package(s) are installed.

This commit also encapsulates the tidy_stats.mixed example in
\donttest{ }, because it takes more than a minute to run on my
system. This should really be fixed, as it does not conform to the
CRAN policy: "Examples should run for no more than a few seconds each:
they are intended to exemplify to the would-be user how to use the
functions in the package. "

Lastly this commit adds the files generated by usethat::use_pipe() to
make pipes in the examples work.